agent-service/pyproject.toml
[project]
name = "pernect-agent-service"
version = "0.1.0"
description = "Pernect AI Agent Service - Multi-agent personality analysis"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.7",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
